refactor(dashboard): drop dead empty-state forwarder, point callers at feedback-state direct
`components/common/empty-state.ts` was a 1-line forwarder:

    export { EmptyState } from './feedback-state'

It produced the same dual-path drift as the `mission-utils` hub
removed in iter#38 — 31 component files imported `EmptyState` through
the forwarder, 55 files imported its siblings (`LoadingState`,
`ErrorState`, etc.) from `./feedback-state` directly, and three test
files (`verification-requests-panel.test.ts`, `tlc-results-panel.test.ts`,
`board/post-detail.test.ts`) had `vi.mock('./common/empty-state', …)`
that would silently no-op the moment a consumer migrated to the direct
path.

Delete the forwarder, repoint all consumers (production + tests + mock
declarations + the sibling import in `mermaid-graph.ts`) at
`./common/feedback-state` directly, and rename the
`'EmptyState (via empty-state re-export)'` describe block in
`empty-state.test.ts` to plain `'EmptyState'` now that the re-export
is gone. The unit test file stays where it is — it covers component
behavior, not the forwarder shape.

Two test files (`verification-requests-panel.test.ts`,
`tlc-results-panel.test.ts`) had a `vi.mock('./common/empty-state', …)`
**and** a separate `vi.mock('./common/feedback-state', …)`. After the
sed-driven path swap they collapsed onto the same path and the second
mock silently overwrote the first — Vitest takes the later one and
drops EmptyState. Merged the two mocks into one factory per file so
EmptyState, ErrorState, and LoadingState are all stubbed in a single
declaration, the way the test was originally written.

iter#38 deleted the mission-utils hub; iter#39 deletes empty-state
which behaved identically — same pattern, same shape, same silent
test-mock risk. Chained dead-code detection continues.
