+++ dashboard/src/components/agent-monitor/live-timeline.ts
@@ -6,7 +6,7 @@ import { signal } from '@preact/signals'
 import { useEffect, useRef, useMemo } from 'preact/hooks'
 import { TimeAgo } from '../common/time-ago'
 import { FilterChips } from '../common/filter-chips'
-import { EmptyState } from '../common/empty-state'
+import { EmptyState } from '../common/feedback-state'
 import { StatusChip, type StatusChipTone } from '../common/status-chip'
 import { journal } from '../../sse'
 import { isErrorJournalEntry } from '../../journal-entry'
