+++ dashboard/src/components/board/board-surface.ts
@@ -6,7 +6,7 @@ import { Card } from '../common/card'
 import { TimeAgo } from '../common/time-ago'
 import { showToast } from '../common/toast'
 import { requestConfirm } from '../common/confirm-dialog'
-import { EmptyState } from '../common/empty-state'
+import { EmptyState } from '../common/feedback-state'
 import { LoadingState } from '../common/feedback-state'
 import { TextInput } from '../common/input'
 import { Select } from '../common/select'
