+++ dashboard/src/components/board/message-room-timeline.ts
@@ -2,7 +2,7 @@ import { html } from 'htm/preact'
 import { useMemo, useState } from 'preact/hooks'
 import { ArrowLeft, AtSign, Braces } from 'lucide-preact'
 import { ActionButton } from '../common/button'
-import { EmptyState } from '../common/empty-state'
+import { EmptyState } from '../common/feedback-state'
 import { RichContent } from '../common/rich-content'
 import { TimeAgo } from '../common/time-ago'
 import { stripStateBlocks } from '../../keeper-message'
