+++ dashboard/src/components/board/post-detail.ts
@@ -5,7 +5,7 @@ import { ActionButton } from '../common/button'
 import { Card } from '../common/card'
 import { TimeAgo } from '../common/time-ago'
 import { showToast } from '../common/toast'
-import { EmptyState } from '../common/empty-state'
+import { EmptyState } from '../common/feedback-state'
 import { LoadingState } from '../common/feedback-state'
 import { RichComposer } from '../common/rich-composer'
 import { RichContent } from '../common/rich-content'
