--- dashboard/src/components/common/empty-state.ts
@@ -1,1 +1,0 @@
-export { EmptyState } from './feedback-state'
