+++ dashboard/src/components/goals/task-detail-overlay.ts
@@ -6,7 +6,7 @@ import { useRef } from 'preact/hooks'
 import { Check, X, ArrowRight, Dot, UserPlus } from 'lucide-preact'
 import { DialogOverlay } from '../common/dialog'
 import { StatusBadge } from '../common/status-badge'
-import { EmptyState } from '../common/empty-state'
+import { EmptyState } from '../common/feedback-state'
 import { ErrorState, LoadingState } from '../common/feedback-state'
 import { RichContent } from '../common/rich-content'
 import { TextInput } from '../common/input'
