+++ dashboard/src/components/governance-monitor.ts
@@ -3,7 +3,7 @@ import { useEffect, useMemo } from 'preact/hooks'
 import { useSignal } from '@preact/signals'
 import { ActionButton } from './common/button'
 import { Card } from './common/card'
-import { EmptyState } from './common/empty-state'
+import { EmptyState } from './common/feedback-state'
 import { ErrorState, LoadingState } from './common/feedback-state'
 import { Select } from './common/select'
 import { TextInput } from './common/input'
