+++ dashboard/src/components/keeper-cognition-inspector.ts
@@ -2,7 +2,7 @@ import { html } from 'htm/preact'
 import type { Keeper } from '../types'
 import { navigate, route } from '../router'
 import { keepers } from '../store'
-import { EmptyState } from './common/empty-state'
+import { EmptyState } from './common/feedback-state'
 import { FilterChips } from './common/filter-chips'
 import { PanelCard } from './common/panel-card'
 import { KeeperBadge } from './keeper-badge'
