+++ dashboard/src/components/surface-readiness-panel.ts
@@ -7,7 +7,7 @@ import { formatTimeAgoEn } from '../lib/format-time'
 import { AsyncContainer } from './common/async-container'
 import { Card } from './common/card'
 import { FilterChips } from './common/filter-chips'
-import { EmptyState } from './common/empty-state'
+import { EmptyState } from './common/feedback-state'
 import { StatusChip, type StatusChipTone } from './common/status-chip'
 import { KpiStripIsland, type KpiStripIslandData } from './kpi-strip-island'
 import { asBoolean, asRecordArray, asString, isRecord } from './common/normalize'
