+++ dashboard/src/components/tools/tool-full-inventory.ts
@@ -4,7 +4,7 @@ import { html } from 'htm/preact'
 import { useEffect, useRef, useCallback } from 'preact/hooks'
 import { ActionButton } from '../common/button'
 import { VirtualList } from '../common/virtual-list'
-import { EmptyState } from '../common/empty-state'
+import { EmptyState } from '../common/feedback-state'
 import { ErrorState } from '../common/feedback-state'
 import { TextInput } from '../common/input'
 import { Select } from '../common/select'
