+++ dashboard/src/components/verification-requests-panel.ts
@@ -24,7 +24,7 @@ import {
 } from '../api/dashboard'
 import { Btn } from './btn'
 import { Card } from './common/card'
-import { EmptyState } from './common/empty-state'
+import { EmptyState } from './common/feedback-state'
 import { ErrorState, LoadingState } from './common/feedback-state'
 import { StatusChip } from './common/status-chip'
 import { relativeTime } from '../lib/format-time'
