+++ dashboard/src/components/verification-specs-panel.ts
@@ -20,7 +20,7 @@ import {
 } from '../api/dashboard'
 import { Btn } from './btn'
 import { Card } from './common/card'
-import { EmptyState } from './common/empty-state'
+import { EmptyState } from './common/feedback-state'
 import { ErrorState, LoadingState } from './common/feedback-state'
 import { StatusChip } from './common/status-chip'
 import { FilterChips } from './common/filter-chips'
